add session and authentication

+++ routes/user.js
@@ -3,7 +3,7 @@ var router = express.Router();
 
 const UserLogic = require('../modules/logic/userlogic')
 const ApplicationUserLogic = require('../modules/logic/applicationuserlogic')
-
+const Formatter = require('../modules/util/formatter')
 
 router.post('/register', function (req, res){
     let user = req.body;
@@ -20,18 +20,8 @@ router.post('/register', function (req, res){
 
     UserLogic.register(user).then(function (saveduser)
     {
+        saveduser = Formatter.removeXSS(saveduser);
         res.send(saveduser);
-    }).catch(function (err){
-        console.log("error")
-        res.send(err);
-    })
-})
-
-router.get('/', function (req, res){
-
-    UserLogic.findAll().then(function (users)
-    {
-        res.send(users);
     }).catch(function (err){
         console.log("error")
         console.log(err);
@@ -39,29 +29,68 @@ router.get('/', function (req, res){
     })
 })
 
+router.get('/', function (req, res){
+
+  let appSession = req.session;
+  if(appSession != null && appSession.loggedInUser != null && appSession.loggedInUser.isadmin != null)
+  {
+    UserLogic.findAll().then(function (users)
+    {
+        users = Formatter.removeXSS(users);
+        res.send(users);
+    }).catch(function (err){
+        console.log("error")
+        console.log(err);
+        res.send(err);
+    })
+  }
+  else
+  {
+    res.send("<b>404</b>");
+  }
+})
+
 
 router.get('/get/:id', function (req, res){
     let id = req.params.id;
-  
-    UserLogic.get(id).then(function (user)
+
+    let appSession = req.session;
+    if(appSession != null && appSession.loggedInUser != null && appSession.loggedInUser.isadmin != null)
     {
-      res.send(user);
-    }).catch(function (err){
-      console.log("error")
-      res.send(err);
-    })
+      UserLogic.get(id).then(function (user)
+      {
+        user = Formatter.removeXSS(user);
+        res.send(user);
+      }).catch(function (err){
+        console.log("error")
+        res.send(err);
+      })
+    }
+    else
+    {
+      res.send("<b>404</b>");
+    }
 })
 
 router.get('/search/:keyword', function (req, res){
     let keyword = req.params.keyword;
-  
-    UserLogic.findByKeyword(keyword).then(function (users)
+
+    let appSession = req.session;
+    if(appSession != null && appSession.loggedInUser != null && appSession.loggedInUser.isadmin != null)
     {
-      res.send(users);
-    }).catch(function (err){
-      console.log("error")
-      res.send(err);
-    })
+      UserLogic.findByKeyword(keyword).then(function (users)
+      {
+        users = Formatter.removeXSS(users);
+        res.send(users);
+      }).catch(function (err){
+        console.log("error")
+        res.send(err);
+      })
+    }
+    else
+    {
+      res.send("<b>404</b>");
+    }
 })
 
 router.post('/login', function (req, res){
@@ -69,6 +98,7 @@ router.post('/login', function (req, res){
   
     UserLogic.login(user.email, user.password).then(function (saveduser)
     {
+      saveduser = Formatter.removeXSS(saveduser);
       res.send(saveduser);
     }).catch(function (err){
       console.log("error")
@@ -77,17 +107,20 @@ router.post('/login', function (req, res){
 })
 
 router.post('/loginbyapp', function (req, res){
-  let user = req.body;
-  let appSession = req.session;
+    let user = req.body;
+    let appSession = req.session;
+    UserLogic.loginByApp(user.email, user.password, user.appId).then(function (saveduser)
+    {
+      appSession.loggedInUser = saveduser.payload;
+      saveduser = Formatter.removeXSS(saveduser);
+      res.send(saveduser);
 
-  UserLogic.loginByApp(user.email, user.password, user.appId).then(function (saveduser)
-  {
-    appSession.loggedInUser = saveduser.payload;
-    res.send(saveduser);
-  }).catch(function (err){
-    console.log("error")
-    res.send(err);
-  })
+    }).catch(function (err){
+      console.log("error")
+      console.log(err);
+      res.send(err);
+    })
+
 })
 
 
@@ -99,6 +132,7 @@ router.post('/loginbyemail', function (req, res){
   UserLogic.loginByEmail(user.email).then(function (saveduser)
   {
     appSession.loggedInUser = saveduser.payload;
+    saveduser = Formatter.removeXSS(saveduser);
     res.send(saveduser);
   }).catch(function (err){
     console.log("error")
@@ -109,33 +143,49 @@ router.post('/loginbyemail', function (req, res){
 router.post('/update/:id', function (req, res){
   let user = req.body;
   let id = req.params.id;
-
-  UserLogic.update(id, user).then(function (saveduser)
+  let appSession = req.session;
+  if(appSession != null && appSession.loggedInUser != null && appSession.loggedInUser.isadmin != null)
   {
-    res.send(saveduser);
-  }).catch(function (err){
-    console.log("error")
-    res.send(err);
-  })
+    UserLogic.update(id, user).then(function (saveduser)
+    {
+      saveduser = Formatter.removeXSS(saveduser);
+      res.send(saveduser);
+    }).catch(function (err){
+      console.log("error")
+      res.send(err);
+    })
+  }
+  else
+  {
+    res.send("<b>404</b>");
+  }
 })
 
 router.get('/delete/:id', function (req, res){
   let id = req.params.id;
-
-  ApplicationUserLogic.deleteByUserId(id).then(function (response){
-    UserLogic.delete(id).then(function (result)
-    {
-      res.send(result);
+  let appSession = req.session;
+  if(appSession != null && appSession.loggedInUser != null && appSession.loggedInUser.isadmin != null)
+  {
+    ApplicationUserLogic.deleteByUserId(id).then(function (response){
+      UserLogic.delete(id).then(function (result)
+      {
+        result = Formatter.removeXSS(result);
+        res.send(result);
+      }).catch(function (err){
+        console.log("error")
+        console.log(err)
+        res.send(JSON.stringify(err));
+      })
     }).catch(function (err){
       console.log("error")
       console.log(err)
       res.send(JSON.stringify(err));
     })
-  }).catch(function (err){
-    console.log("error")
-    console.log(err)
-    res.send(JSON.stringify(err));
-  })
+  }
+  else
+  {
+    res.send("<b>404</b>");
+  }
 
 
 })
@@ -144,6 +194,7 @@ router.get('/delete/:id', function (req, res){
 router.get('/session/:sessionid', function (req, res){
   let sessionId = req.params.sessionid;
   UserLogic.checkSession(sessionId).then(function (result){
+    result = Formatter.removeXSS(result);
     res.send(result);
   }).catch(function(error)
   {
